Update 2014_10_13_000001_create_users_table.php
business_name to users_business_name

unless I'm missing something, when I was working with joined tables the command to get something from another tables was $selectedBrands->brand_name, this would be an issue if selected Brands was joining users and brands and both only had name. $selectedBrands->name would be open to the system to figure out what we meant. I really feel this isn't how it's supposed to work, but that is the way it seems to be working.